+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Column "reference" on Form table of the database structure
 - Column "reference" on FormQuestion table of the database structure
 - Fields "reference" on forms and questions on oxe-web-admin
+- "TEXTAREA" question type on forms
 
 ## [1.13.6] - 2022-10-12